Tweak restaurant.edit
In a similar fashion as restaurant.show, by default map and the rest of
the form are in their own rows but when SM breakpoint is reached, they
are on the same line.

Map-block component is tweaked so that the map inside the block can get
attributes.

On SM breakpoint map block and form are the same height, description
fills the remaining space.