Handle locale gracefully (#19)
* Handle wrong locale in URL gracefully

Redirect to the default language page and send a flash (the message is
hardcoded in English though).

* Create bare component folder for organization

The component folder starts to become a bit cluttered and this is why
"bare" folder appeared. In future there would probably be much more
icons so it's a reasonable step to make.

* Respect the target URL in EnsureLanguage

thisWithLocale is extracted to the helper file and is now used in
EnsureLanguage to not throw users off the page they intended to visit.